feat(bagdes-slider): enable slider for desktop
