+++ tsconfig.json
@@ -12,6 +12,7 @@
     "baseUrl": "./",
     "paths": {
       "components/*": ["src/components/*"],
+      "utils/*": ["src/utils/*"],
       "sections/*": ["src/sections/*"],
       "constants/*": ["src/constants/*"]
     }
